requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
jinja2==3.1.4
pyyaml==6.0.2
anthropic==0.43.0
openai==1.54.0
google-genai==0.3.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx>=0.27,<0.28

